Renamed anki.utility
