+++ TODO.md
@@ -20,5 +20,5 @@
 + ~~rename `Vehicle.__notify_handler__`~~
 + ~~rename all `[...]Exception`s to `[...]Error`s~~
     ~~Implement deprecated aliases here as well~~ (Implemented as regular aliases)
-+ rename `anki.utility` and `anki.utility.util` since they don't actually provide any
++ ~~rename `anki.utility` and `anki.utility.util` since they don't actually provide any~~
 + ~~Change the `raise` usages to `raise from`~~
