+++ anki/misc/deprecated_alias/__main__.py
